Split property-based smoke and scheduled suites
