+++ .github/workflows/ci.yml
@@ -44,7 +44,7 @@ jobs:
     - name: Build Example
       run: dotnet build --configuration Release --framework ${{ matrix.target-framework }} ConsoleMarkdownRenderer.Example
     - name: Test Specs
-      run: dotnet test --configuration Release --framework ${{ matrix.target-framework }} ConsoleMarkdownRenderer.Spectre.Tests
+      run: dotnet test --configuration Release --framework ${{ matrix.target-framework }} ConsoleMarkdownRenderer.Spectre.Tests --filter 'TestCategory!=ScheduledFuzz'
     - name: Test
       run: dotnet test --configuration Release --framework ${{ matrix.target-framework }} ConsoleMarkdownRenderer.Tests
     - name: Test Fakes
